requirements.txt
Django==6.0.4
Pillow==11.2.1
pandas==2.2.3
numpy==2.2.5
python-decouple==3.8
whitenoise==6.9.0
django-allauth[google]==65.16.1
PyJWT==2.12.1
cryptography==48.0.0
midtransclient==1.4.2
